Kindleでファイル未添付となる問題を修正
