+++ src/lib.rs
@@ -1,5 +1,4 @@
 extern crate lettre;
 
-pub mod mailer;
 pub mod subcommand;
 pub mod config;
